+++ changelog.md
@@ -27,5 +27,7 @@
 
 //Pipes
 22. Add built-in Pipes for Date, Price and Name for cart-item
+23. Change getProducts() in Product Service to return Observable,
+add getProducts2 ro return Promise, use AsyncPipe to display result at product-list component